Animation works as intended, but glitchy
The shimmer animation now runs once when .dino-img:hover and resets
when no hover.

But there is a noticeable color shift in the background when the
animation is activated. It would seem the no hover condition is
starting the background linear-gradient at a lighter position than 0%,
so the background appears lighter. Until img:hover, where the color
shift becomes slightly darker as it starts from the intended background
color.

Fix this.