Create button over img, then comment it out
Placed a transparent button over .dino-img in an attempt to make a
trigger to activate the Shimmer animation for one full iteration
repeatedly with each click. Most solutions I found rely on js, which
is beyond my current knowledge base.

I tried adding a checkbox input to the button to act as a toggle that
activated the animation when clicked, checked or unchecked. Got over
my head and decided to revert back to the hover trigger.

Commented out the button html and css, as it may come in handy later.