Fix opacity fade in/out and animation trigger
Opacity transition on .chick-img is the fade out, and the active class
is the fade in. Duh. I had them switched.

Sorta of the same thing with the shimmer animation. Activating will
now reset the animation, and not:active will play the entire animation
when the mouse is released (and when the page loads and reloads).

Whoop whoop.